feat(symbolic-go): Basic skeleton wrapper (#1)
## Which problem is this PR solving?

A basic skeleton of a wrapper for the symbolic rust library in Go.

## Short description of the changes

Builds the symbolic library and invokes it using cgo. Runs a bunch of
source map tests against this code to ensure it does what it says it
does.

## How to verify that this has the expected result

Check the test output, it builds and compiles and runs the tests in
GitHub actions currently.